Recalculate poll period after role change to child.
If we configure CSL parameters before a command to attach, then the child has configured valid CSL state, but when trying to configure the auto-sync CSL poll timeout, the check to IsCslEnabled->IsCslCapable fails because IsChild() fails.

When handling the child ID response success, we now trigger poll-sender's recalculation, and rx-on-idle states, after role change to child.