[android] fix getting version errors
When adding the openthread to master branch of the manifest, the
Presubmit build generates an error on branch git_master on target
aosp_crosshatch-userdebug. The error shows that the openthread source
code is not under a git repo. So that we can't use the `git describe`
command to get the version string.

This CL checks if the source code is under a git repo, if not, retuns
a string "Unknown" as the version string.

Bug: 203492771
Test: build
Change-Id: I635544c6f46fed4db2df4948af3eea3f42d8b770