[mlr-manager] add short delay for processing `Ip6MulticastSubscribed` event
This commit updates `MlrManager` such that it adds a short wait delay
before taking action when an IPv6 multicast address is added, i.e., it
processes a `kEventIp6MulticastSubscribed` event from `Notifier` and
calls `UpdateLocalSubscriptions()`. The delay is added only when the
`mDelayTimer` is not already running (i.e. on the first event). This
allows user to add multiple addresses back-to-back and ensure that
they are processed together.