[cli] support 1 or more mcast addresses for `ipmaddr add` command (#7253)
Because the 2 addresses are registered sequentially from the same
method call, the OpenThread stack will not have had the opportunity to
send out MLR.req with a single address. So this should result in
sending out one MLR.req with two addresses inside, if 2 addresses are
provided.  Without changes in the APIs.