[posix] resolve the RCP timeout issue
When calling `scan energy` command continuously, the SPI based RCP Host
crashes and the log outputs RCP timeout error message. The root cause is
that the interval between two Spinel commands is very small. After
processing the previous SPI packet, the RCP needs to takes a certain amount
of time to re-set the TX/RX buffer of the SPI slave controller. During
the process of RCP setting the SPI slave TX/RX buffers, the Host starts
an SPI transfer, and then the RCP side loses the packet at this time.

This PR checks the result of the SPI transmission when sending the Spinel
command. If the SPI transmission is failed, the Host retransmits the SPI
frame. This PR also changed the fd of the interrupt pin to O_NONBLOCK mode.